chore: file structure for pages
